fix wish delete exception and return reponse

+++ server/src/main/java/com/mbafour/mba/service/WishDeleteService.java
@@ -1,6 +1,7 @@
 package com.mbafour.mba.service;
 
 import com.mbafour.mba.domain.repository.WishRepository;
+import com.mbafour.mba.exception.NoExistWishIdException;
 import org.springframework.stereotype.Service;
 import lombok.AllArgsConstructor;
 
@@ -10,7 +11,13 @@ public class WishDeleteService {
 
     private final WishRepository wishRepository;
 
-    public  void deleteWish(Long wishId) {
+    public  void deleteWish(Long wishId) throws Exception{
+
+        if(!wishRepository.existsById(wishId)) {
+            throw new NoExistWishIdException();
+        }
+
+
         wishRepository.deleteById(wishId);
    }
